+++ tests/quality_of_life/QOL_testing_helpers.py
@@ -1,9 +1,6 @@
-"""Test that all indicator accessor functions return dataframe. Had issues with some
-returning series"""
-import pytest
-import pandas as pd
 from aggregate.quality_of_life.access_to_jobs import access_to_jobs
 from aggregate.quality_of_life.access_to_open_space import park_access
+from aggregate.quality_of_life.access_transit import access_subway_and_access_ADA
 from aggregate.quality_of_life.covid_death import covid_death
 from aggregate.quality_of_life.education_outcome import get_education_outcome
 from aggregate.quality_of_life.heat_vulnerability import load_clean_heat_vulnerability
@@ -16,10 +13,5 @@ accessors = [
     access_to_jobs,
     covid_death,
     load_clean_heat_vulnerability,
+    access_subway_and_access_ADA,
 ]
-
-
-@pytest.mark.parametrize("accessor", accessors)
-@pytest.mark.parametrize("geography", ["puma", "borough", "citywide"])
-def test_rv_dataframe(accessor, geography):
-    assert isinstance(accessor(geography), pd.DataFrame)
